create new component home & set new routes
