+++ portal/package.json
@@ -11,7 +11,7 @@
     "react": "^18.2.0",
     "react-chartjs-2": "3.1.1",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.14.2",
+    "react-router-dom": "^6.20.1",
     "react-scripts": "5.0.1",
     "react-select": "^5.7.4",
     "react-toastify": "^9.1.3",
