+++ portal/src/app/components/home/Home.module.scss
@@ -1,0 +1,49 @@
+@import "src/styles/variables-keys";
+@import "src/styles/z-index";
+
+.app_wrapper {
+  padding: 80px;
+}
+
+.init_state_wrapper {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  margin-block-start: 120px;
+  font-family: var($fontBold);
+}
+
+.app_spinner_overlay {
+  inline-size: 100%;
+  block-size: 100%;
+  position: absolute;
+  background: var($backgroundColorWhite);
+  opacity: 0.6;
+  top: 0;
+  left: 0;
+  z-index: $appSpinnerZindex;
+}
+
+.app_spinner {
+  position: sticky;
+  top: 50%;
+  left: 50%;
+  text-align: center;
+}
+
+.protocol_query_title {
+  font-size: 18px;
+  font-family: var($fontMedium);
+  margin-block-end: 40px;
+}
+
+.response_wrapper {
+  margin-block-start: 60px;
+}
+
+.response_link {
+  border-radius: 12px;
+  background-color: var($attPurple);
+  padding: 14px;
+}
