+++ portal/src/app/components/home/index.ts
@@ -1,0 +1,1 @@
+export * from './Home';
